% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_chol.R
\name{log_chol}
\alias{log_chol}
\alias{log_chol_inv}
\title{Computes the Log Cholesky Decomposition and the Inverse}
\usage{
log_chol(x)

log_chol_inv(x)
}
\arguments{
\item{x}{A positive definite matrix or a vector with a log Cholesky
decomposition.}
}
\description{
Computes the Log Cholesky Decomposition and the Inverse
}
\examples{
set.seed(1)
S <- drop(rWishart(1, 10, diag(10)))
stopifnot(isTRUE(all.equal(S, log_chol_inv(log_chol(S)))))

}
