% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif.R
\name{mmcif_data}
\alias{mmcif_data}
\title{Setup Object to Compute the Log Composite Likelihood}
\usage{
mmcif_data(
  formula,
  data,
  cause,
  time,
  cluster_id,
  max_time,
  spline_df = 3L,
  left_trunc = NULL,
  ghq_data = NULL,
  strata = NULL,
  knots = NULL
)
}
\arguments{
\item{formula}{\code{formula} for covariates in the risk and trajectories.}

\item{data}{\code{data.frame} with the covariate and outcome information.}

\item{cause}{an integer vector with the cause of each outcome. If there are
\code{n_causes} of outcome, then the vector should have values in
\code{1:(n_causes + 1)} with \code{n_causes + 1} indicating censoring.}

\item{time}{a numeric vector with the observed times.}

\item{cluster_id}{an integer vector with the cluster id of each individual.}

\item{max_time}{the maximum time after which there are no observed events. It
is denoted by \eqn{\delta} in the original article.}

\item{spline_df}{degrees of freedom to use for each spline in the
cumulative incidence functions.}

\item{left_trunc}{numeric vector with left-truncation times. \code{NULL}
implies that there is not any individuals with left-truncation.}

\item{ghq_data}{the default Gauss-Hermite quadrature nodes and weights to
use. It should be stored as a list with two elements called \code{"node"}
and \code{"weight"}. A default is provided if \code{NULL} is passed.}

\item{strata}{an integer vector or a factor vector with the strata of each
individual. \code{NULL} implies that there are no strata.}

\item{knots}{A list of lists with knots for the splines. The innner lists
 needs to have elements called \code{"knots"} and
\code{"boundary_knots"} which are passed to a function like \code{\link{ns}}.
\code{NULL} yields defaults based on the quantiles of the observed event
times. Note that the knots needs to be on the
\code{atanh((time - max_time / 2) / (max_time / 2))} scale.}
}
\description{
Setup Object to Compute the Log Composite Likelihood
}
\seealso{
\code{\link{mmcif_fit}}, \code{\link{mmcif_start_values}} and
\code{\link{mmcif_sandwich}}.
}
