% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif.R
\name{mmcif_fit}
\alias{mmcif_fit}
\title{Fits a Mixed Competing Risk Model}
\usage{
mmcif_fit(
  par,
  object,
  n_threads = 1L,
  control.outer = list(itmax = 100L, method = "nlminb", kkt2.check = FALSE, trace =
    FALSE),
  control.optim = list(eval.max = 10000L, iter.max = 10000L),
  ghq_data = object$ghq_data,
  ...
)
}
\arguments{
\item{par}{numeric vector with parameters. This is using a log
Cholesky decomposition for the covariance matrix.}

\item{object}{an object from \code{\link{mmcif_data}}.}

\item{n_threads}{the number of threads to use.}

\item{control.outer, control.optim, ...}{arguments passed to
\code{\link[alabama]{auglag}}.}

\item{ghq_data}{the Gauss-Hermite quadrature nodes and weights to use.
It should be a list with two elements called \code{"node"}. The argument
can also be a list with different number of quadrature nodes. In this case,
fits are successively using the previous fit. This may reduce the computation
time by starting with fewer quadrature nodes.}
}
\description{
Fits a Mixed Competing Risk Model
}
\seealso{
\code{\link{mmcif_data}}, \code{\link{mmcif_start_values}} and
\code{\link{mmcif_sandwich}}.
}
