% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif.R
\name{mmcif_logLik}
\alias{mmcif_logLik}
\alias{mmcif_logLik_grad}
\title{Evaluates the Log Composite Likelihood and its Gradient}
\usage{
mmcif_logLik(
  object,
  par,
  ghq_data = object$ghq_data,
  n_threads = 1L,
  is_log_chol = FALSE
)

mmcif_logLik_grad(
  object,
  par,
  ghq_data = object$ghq_data,
  n_threads = 1L,
  is_log_chol = FALSE
)
}
\arguments{
\item{object}{an object from \code{\link{mmcif_data}}.}

\item{par}{numeric vector with parameters. This is either using a log
Cholesky decomposition for the covariance matrix or the covariance matrix.}

\item{ghq_data}{the Gauss-Hermite quadrature nodes and weights to
use. It should be a list with two elements called \code{"node"}
and \code{"weight"}. A default is provided if \code{NULL} is passed.}

\item{n_threads}{the number of threads to use.}

\item{is_log_chol}{logical for whether a log Cholesky decomposition is used
for the covariance matrix or the full covariance matrix.}
}
\description{
Evaluates the Log Composite Likelihood and its Gradient
}
