% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif-pd.R
\name{mmcif_pd_cond}
\alias{mmcif_pd_cond}
\alias{mmcif_pd_bivariate}
\title{Computes Marginal Measures Using Two Observations}
\usage{
mmcif_pd_cond(
  par,
  object,
  newdata,
  cause,
  time,
  left_trunc = NULL,
  ghq_data = object$ghq_data,
  strata = NULL,
  which_cond,
  type_cond = "derivative",
  type_obs = "cumulative"
)

mmcif_pd_bivariate(
  par,
  object,
  newdata,
  cause,
  time,
  left_trunc = NULL,
  ghq_data = object$ghq_data,
  strata = NULL,
  use_log = FALSE,
  type = c("cumulative", "cumulative")
)
}
\arguments{
\item{par}{numeric vector with the model parameters.}

\item{object}{an object from \code{\link{mmcif_data}}.}

\item{newdata}{a \code{data.frame} with data for the observations. It needs
to have two rows.}

\item{cause}{an integer vector with the cause of each outcome. If there are
\code{n_causes} of outcome, then the vector should have values in
\code{1:(n_causes + 1)} with \code{n_causes + 1} indicating censoring.}

\item{time}{a numeric vector with the observed times.}

\item{left_trunc}{numeric vector with left-truncation times. \code{NULL}
implies that there are not any individuals with left-truncation.}

\item{ghq_data}{the Gauss-Hermite quadrature nodes and weights to
use. It should be a list with two elements called \code{"node"}
and \code{"weight"}. A default is provided if \code{NULL} is passed.}

\item{strata}{an integer vector or a factor vector with the strata of each
individual. \code{NULL} implies that there are no strata.}

\item{which_cond}{an integer with value one or two for the index of the
individual that is being conditioned on.}

\item{type_cond}{a character for the type of outcome that is being
conditioned on.
\code{"derivative"} for the derivative of a CIF or
\code{"cumulative"} for a CIF or the survival probability.}

\item{type_obs}{a character the type of conditional measure. It can be
\code{"derivative"} for the derivative of a CIF,
\code{"cumulative"} for a CIF or the survival probability, and
\code{"hazard"} for the hazard.}

\item{use_log}{a logical for whether the returned output should be on the
log scale.}

\item{type}{a 2D character vector for the type of measures for each
observation. The elements can be
\code{"derivative"} for the derivative of a CIF or
\code{"cumulative"} for a CIF or the survival probability.}
}
\description{
Computes Marginal Measures Using Two Observations
}
\seealso{
\code{\link{mmcif_pd_univariate}} and \code{\link{mmcif_fit}}.
}
