% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif-pd.R
\name{mmcif_pd_univariate}
\alias{mmcif_pd_univariate}
\title{Computes Marginal Measures for One Observation}
\usage{
mmcif_pd_univariate(
  par,
  object,
  newdata,
  cause,
  time,
  left_trunc = NULL,
  ghq_data = object$ghq_data,
  strata = NULL,
  use_log = FALSE,
  type = "cumulative"
)
}
\arguments{
\item{par}{numeric vector with the model parameters.}

\item{object}{an object from \code{\link{mmcif_data}}.}

\item{newdata}{a \code{data.frame} with data for the observation. It needs
to have one row.}

\item{cause}{an integer vector with the cause of each outcome. If there are
\code{n_causes} of outcome, then the vector should have values in
\code{1:(n_causes + 1)} with \code{n_causes + 1} indicating censoring.}

\item{time}{a numeric vector with the observed times.}

\item{left_trunc}{numeric vector with left-truncation times. \code{NULL}
implies that there are not any individuals with left-truncation.}

\item{ghq_data}{the Gauss-Hermite quadrature nodes and weights to
use. It should be a list with two elements called \code{"node"}
and \code{"weight"}. A default is provided if \code{NULL} is passed.}

\item{strata}{an integer vector or a factor vector with the strata of each
individual. \code{NULL} implies that there are no strata.}

\item{use_log}{a logical for whether the returned output should be on the
log scale.}

\item{type}{a character for the type of measures for the observation. It
can have value
\code{"derivative"} for the derivative of a CIF or
\code{"cumulative"} for a CIF or the survival probability.}
}
\description{
Computes Marginal Measures for One Observation
}
\seealso{
\code{\link{mmcif_pd_bivariate}} and \code{\link{mmcif_fit}}.
}
