% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif.R
\name{mmcif_sandwich}
\alias{mmcif_sandwich}
\title{Computes the Sandwich Estimator}
\usage{
mmcif_sandwich(
  object,
  par,
  ghq_data = object$ghq_data,
  n_threads = 1L,
  eps = 0.01,
  scale = 2,
  tol = 1e-08,
  order = 6L
)
}
\arguments{
\item{object}{an object from \code{\link{mmcif_data}}.}

\item{par}{numeric vector with the parameters to compute the sandwich
estimator at.}

\item{ghq_data}{the Gauss-Hermite quadrature nodes and weights to
use. It should be stored as a list with two elements called \code{"node"}
and \code{"weight"}. A default is provided if \code{NULL} is passed.}

\item{n_threads}{the number of threads to use.}

\item{eps}{determines the step size in the numerical differentiation using
\code{max(eps^2, |par[i]| * eps)}.}

\item{scale}{scaling factor in the Richardson extrapolation.}

\item{tol}{relative convergence criteria in the extrapolation given
by \code{max(tol, |g[j]| * tol)} with \code{g} being the gradient.}

\item{order}{maximum number of iteration of the Richardson extrapolation.}
}
\description{
Computes the sandwich estimator of the covariance matrix. The parameter that
is passed is using the log Cholesky decomposition. The Hessian is computed
using numerical differentiation with Richardson extrapolation to refine the
estimate.
}
\seealso{
\code{\link{mmcif_fit}} and \code{\link{mmcif_data}}.
}
