% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif.R
\name{mmcif_start_values}
\alias{mmcif_start_values}
\title{Finds Staring Values}
\usage{
mmcif_start_values(object, n_threads = 1L, vcov_start = NULL)
}
\arguments{
\item{object}{an object from \code{\link{mmcif_data}}.}

\item{n_threads}{the number of threads to use.}

\item{vcov_start}{starting value for the covariance matrix of the random
effects. \code{NULL} yields the identity matrix.}
}
\description{
Finds Staring Values
}
