% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif.R
\name{mmcif_data}
\alias{mmcif_data}
\title{Setup Object to Compute the Log Composite Likelihood}
\usage{
mmcif_data(formula, data, cause, time, cluster_id, max_time)
}
\arguments{
\item{formula}{\code{formula} for covariates in the risk and trajectories.}

\item{data}{\code{data.frame} with the covariate and outcome information.}

\item{cause}{an integer vector with the cause of each outcome. If there are
\code{n_causes} of outcome, then the vector should have values in
\code{1:(n_causes + 1)} with \code{n_causes + 1} indicating censoring.}

\item{time}{a numeric vector with the observed time.}

\item{cluster_id}{an integer vector with the cluster id of each individual.}

\item{max_time}{the maximum time after which there are no observed events. It
is denoted by \eqn{\delta} in the original article.}
}
\description{
Setup Object to Compute the Log Composite Likelihood
}
