% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif.R
\name{mmcif_data}
\alias{mmcif_data}
\title{Setup Object to Compute the Log Composite Likelihood}
\usage{
mmcif_data(
  formula,
  data,
  cause,
  time,
  cluster_id,
  max_time,
  spline_df = 3L,
  left_trunc = NULL,
  ghq_data = NULL
)
}
\arguments{
\item{formula}{\code{formula} for covariates in the risk and trajectories.}

\item{data}{\code{data.frame} with the covariate and outcome information.}

\item{cause}{an integer vector with the cause of each outcome. If there are
\code{n_causes} of outcome, then the vector should have values in
\code{1:(n_causes + 1)} with \code{n_causes + 1} indicating censoring.}

\item{time}{a numeric vector with the observed times.}

\item{cluster_id}{an integer vector with the cluster id of each individual.}

\item{max_time}{the maximum time after which there are no observed events. It
is denoted by \eqn{\delta} in the original article.}

\item{spline_df}{degrees of freedom to use for each spline in the
cumulative incidence functions.}

\item{left_trunc}{numeric vector with left-truncation times. \code{NULL}
implies that there is not any individuals with left-truncation.}

\item{ghq_data}{the default Gauss-Hermite quadrature nodes and weights to
use. It should be stored as a list with two elements called \code{"node"}
and \code{"weigth"}. A default is provided if \code{NULL} is passed.}
}
\description{
Setup Object to Compute the Log Composite Likelihood
}
