% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif.R
\name{mmcif_fit}
\alias{mmcif_fit}
\title{Fits a Mixed Competing Risk Model}
\usage{
mmcif_fit(
  par,
  object,
  n_threads = 1L,
  control = list(maxit = 10000L, reltol = 1e-10),
  method = "BFGS",
  outer.eps = 1e-07,
  ...
)
}
\arguments{
\item{par}{numeric vector with parameters. This is using a log
Cholesky decomposition for the covariance matrix.}

\item{object}{an object from \code{\link{mmcif_data}}.}

\item{n_threads}{the number of threads to use.}

\item{control, method, outer.eps, ...}{arguments passed to
\code{\link{constrOptim}}.}
}
\description{
Fits a Mixed Competing Risk Model
}
\seealso{
\code{\link{mmcif_data}}, \code{\link{mmcif_start_values}} and
\code{\link{mmcif_sandwich}}.
}
