% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcif.R
\name{mmcif_sandwich}
\alias{mmcif_sandwich}
\title{Computes the Sandwich Estimator}
\usage{
mmcif_sandwich(
  object,
  par,
  ghq_data = object$ghq_data,
  n_threads = 1L,
  eps = 1e-04,
  scale = 2,
  tol = 1e-08,
  order = 6L
)
}
\arguments{
\item{object}{an object from \code{\link{mmcif_data}}.}

\item{par}{numeric vector with the parameters to compute the sandwich
estimator at.}

\item{n_threads}{the number of threads to use.}

\item{eps}{determines the step size in the numerical differentiation using
\code{max(eps, |par[i]| * eps)}.}

\item{order}{maximum number of iteration of the Richardson extrapolation.}

\item{scaling}{scaling factor in the Richardson extrapolation.}

\item{relative}{convergence criteria in the extrapolation given
by \code{max(tol, |g[j]| * tol)} with \code{g} being the gradient.}
}
\description{
Computes the sandwich estimator of the covariance matrix. The parameter that
is passed is using the log Cholesky decomposition. The Hessian is computed
using numerical differentiation with Richardson extrapolation to refine the
estimate.
}
